move folder-related code from psynclib to pfolder
